Replace LOGGER by LOG
Replace LOGGER by LOG to follow the 
OpenDaylight recommendations [1].

[1]
https://wiki.opendaylight.org/view/BestPractices/Logging_Best_Practices

+++ vpnservice/bgpmanager/bgpmanager-impl/src/main/java/org/opendaylight/netvirt/bgpmanager/VtyshCli.java
@@ -21,7 +21,6 @@ import org.apache.karaf.shell.console.OsgiCommandSupport;
 
 @Command(scope = "odl", name = "show-bgp", description = "")
 public class VtyshCli extends OsgiCommandSupport {
-    // public class DisplayBgpConfigCli {
 
     @Option(name = "--cmd", description = "command to run", required = true, multiValued = false)
     String cmd = null;
@@ -175,8 +174,8 @@ public class VtyshCli extends OsgiCommandSupport {
                 socket.close();
                 return;
             }
-            if ((ip == (int) gt) || (ip == (int) hashChar)) {
-                if (ip == (int) gt) {
+            if (ip == gt || ip == hashChar) {
+                if (ip == gt) {
                     sb.append(gt);
                 } else {
                     sb.append(hashChar);
@@ -207,7 +206,7 @@ public class VtyshCli extends OsgiCommandSupport {
                 socket.close();
                 return;
             }
-            if ((ip == (int) gt) || (ip == (int) hashChar)) {
+            if (ip == gt || ip == hashChar) {
                 break;
             } else if (ip == -1) {
                 session.getConsole().println(sb.toString());
@@ -271,7 +270,5 @@ public class VtyshCli extends OsgiCommandSupport {
             session.getConsole().println(errorMsg);
         }
         socket.close();
-
     }
-
 }
